Add program takes arguments

+++ assembly-add/main.c
@@ -1,8 +1,9 @@
 #include <stdio.h>
 #include <stdlib.h>
+#include <stdint.h>
 #include <string.h>
 
-extern int add(int a, int b);
+extern int32_t add(int32_t a, int32_t b);
 
 int main(int argc, char *argv[]) {
     printf("%s\n", "Adding with assembly!");
@@ -10,8 +11,8 @@ int main(int argc, char *argv[]) {
         printf("Specify two arguments to add\n");
         return 0;
     }
-    int a = atoi(argv[1]);
-    int b = atoi(argv[2]);
+    int32_t a = (int32_t) atoi(argv[1]);
+    int32_t b = (int32_t) atoi(argv[2]);
     printf("%i + %i = %i\n", a, b, add(a, b));
 }
